TITANIC internal refactor for homing, see if anything broke more work on mp sync for projectiles that attach to other things

+++ Projectiles/BossWeapons/BlenderOrbital.cs
@@ -55,7 +55,7 @@ namespace FargowiltasSouls.Projectiles.BossWeapons
             int byUUID = FargoSoulsUtil.GetByUUIDReal(projectile.owner, (int)projectile.localAI[0], ModContent.ProjectileType<BlenderYoyoProj>());
             if (byUUID == -1)
             {
-                if (projectile.owner == Main.myPlayer)
+                if (projectile.owner == Main.myPlayer && projectile.rotation > 0)
                 {
                     projectile.Kill();
                     return;
